requirements.txt
google-cloud-bigquery>=2.0.0
google-cloud-storage>=1.42.3
pandas
openpyxl
pyspark==3.2.0
functions-framework==3.2.0
fsspec
gcsfs
pyarrow
db-dtypes
pyarrow


